# api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_mirastacklabs-ai_mirador-core_internal_models.KPIDefinition:
    properties:
      aggregationWindowHint:
        description: AggregationWindowHint suggests a preferred aggregation window
          (e.g. "1m", "5m").
        type: string
      businessImpact:
        description: BusinessImpact explains the user/business consequence when this
          KPI degrades.
        type: string
      category:
        description: Category is a free-form category for additional grouping/classification.
        type: string
      classifier:
        description: Classifier is the measurement category (latency, errors, tps,
          cpu_utilization, anomaly_score, ...).
        type: string
      componentType:
        description: ComponentType indicates the type of component this KPI maps to
          (springboot, kafka-broker, cassandra-node, node, valkey, ...).
        type: string
      createdAt:
        type: string
      datastore:
        description: Datastore is the telemetry/metrics store where this KPI is sourced
          from (victoriametrics, clickhouse, ...).
        type: string
      definition:
        description: Definition of what the signal means
        type: string
      dimensionsHint:
        description: DimensionsHint lists key dimensions expected on this KPI (e.g.
          ["service.name", "orgId"]).
        items:
          type: string
        type: array
      domain:
        description: Domain indicates the business or technical domain this KPI applies
          to (payments, kafka, cassandra, infra, ...).
        type: string
      emotionalImpact:
        description: EmotionalImpact is an optional short severity/emotive hint for
          narrative generation.
        type: string
      examples:
        description: Examples contains example values/contexts; stored as arbitrary
          JSON objects.
        items:
          additionalProperties: true
          type: object
        type: array
      format:
        type: string
      formula:
        description: Formula holds the raw query/formula string; treated as opaque
          by the engines.
        type: string
      id:
        type: string
      kind:
        description: '"business" or "tech"'
        type: string
      layer:
        description: 'Layer indicates whether the KPI is an impact or cause signal.
          Allowed: "impact", "cause".'
        type: string
      name:
        type: string
      namespace:
        description: |-
          Namespace groups related KPIs (e.g. file or collection name).
          Example: "apigw_springboot_top_10_kpi".
        type: string
      query:
        additionalProperties: true
        description: JSON query definition
        type: object
      queryType:
        description: QueryType indicates the query language or type (MetricsQL, SQL,
          PromQL, etc.).
        type: string
      retryAllowed:
        description: RetryAllowed indicates whether automated retry logic is permitted
          for incidents related to this KPI.
        type: boolean
      sentiment:
        description: '"NEGATIVE", "POSITIVE", or "NEUTRAL" - increase sentiment'
        type: string
      serviceFamily:
        description: ServiceFamily groups related services (apigw, oltp, issuer-bank,
          ...).
        type: string
      signalType:
        description: SignalType is the high-level signal kind (metrics, traces, logs,
          business, synthetic, ...).
        type: string
      source:
        description: Source identifies where this KPI definition originated (seed
          file, tool, etc.).
        type: string
      sourceId:
        description: SourceID is a short identifier within the source (e.g. CSV row
          key).
        type: string
      sparkline:
        additionalProperties: true
        description: JSON sparkline config
        type: object
      tags:
        items:
          type: string
        type: array
      thresholds:
        description: JSON thresholds array
        items:
          $ref: '#/definitions/github_com_mirastacklabs-ai_mirador-core_internal_models.Threshold'
        type: array
      unit:
        type: string
      updatedAt:
        type: string
      visibility:
        description: '"private", "team", "org"'
        type: string
    type: object
  github_com_mirastacklabs-ai_mirador-core_internal_models.KPIDefinitionRequest:
    properties:
      kpiDefinition:
        $ref: '#/definitions/github_com_mirastacklabs-ai_mirador-core_internal_models.KPIDefinition'
    type: object
  github_com_mirastacklabs-ai_mirador-core_internal_models.KPIListResponse:
    properties:
      kpiDefinitions:
        items:
          $ref: '#/definitions/github_com_mirastacklabs-ai_mirador-core_internal_models.KPIDefinition'
        type: array
      nextOffset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_mirastacklabs-ai_mirador-core_internal_models.Threshold:
    properties:
      color:
        description: Display color
        type: string
      description:
        description: Description of the threshold
        type: string
      level:
        description: e.g., "warning", "critical"
        type: string
      operator:
        description: e.g., "gt", "lt", "eq"
        type: string
      value:
        description: Threshold value
        type: number
    type: object
...
host: localhost:8010
info:
  contact:
    email: support@mirastacklabs-ai.org
    name: Platform Builds Team
    url: https://github.com/mirastacklabs-ai/mirador-core
  description: Mirador Core is a comprehensive observability and analytics platform
    that provides KPI definitions and unified queries for monitoring and analyzing
    system metrics.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mirador Core API
  version: 8.0.0
paths:
  /api/v1/kpi/defs:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of KPI definitions with optional filtering
        by tags
      parameters:
      - collectionFormat: csv
        description: Filter by tags (comma-separated)
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: 'Maximum number of results (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: 'Pagination offset (default: 0)'
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mirastacklabs-ai_mirador-core_internal_models.KPIListResponse'
        "400":
          description: 'error: invalid query parameters'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: failed to list KPIs'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get KPI definitions
      tags:
      - kpi-definitions
    post:
      consumes:
      - application/json
      description: Create a new KPI definition or update an existing one. If ID is
        not provided, a new UUID will be generated.
      parameters:
      - description: KPI definition payload
        in: body
        name: kpi
        required: true
        schema:
          $ref: '#/definitions/github_com_mirastacklabs-ai_mirador-core_internal_models.KPIDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: ok, id: kpi_id'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid payload or validation error'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: failed to upsert KPI'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update KPI definition
      tags:
      - kpi-definitions
  /api/v1/kpi/defs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a KPI definition by its ID. Requires confirmation via query
        parameter.
      parameters:
      - description: KPI definition ID
        in: path
        name: id
        required: true
        type: string
      - description: Confirmation flag (1, true, or yes)
        enum:
        - "1"
        - "true"
        - "yes"
        in: query
        name: confirm
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: deleted'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: missing id or confirmation required'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: failed to delete KPI'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete KPI definition
      tags:
      - kpi-definitions
...
swagger: "2.0"